+++ lib/l10n/supported_languages.dart
@@ -19,5 +19,6 @@ final Map<Locale, (String, String)> supportedLanguages = {
   const Locale("uk", "UA"): ("Ukrainian (Ukraine)", "Українська (Україна)"),
   const Locale("ar"): ("Arabic", "العربية"),
   const Locale("fa", "IR"): ("Persian (Iran)", "فارسی (ایران)"),
+  const Locale("zh", "CN"): ("Chinese (Simplified, China Mainland)", "简体中文 (中国大陆)"),
   const Locale("zh", "TW"): ("Chinese (Traditional, Taiwan)", "正體中文 (台灣)"),
 };